script/lib/distro.rb: add RHEL/Rocky 10 platform
Rocky Linux 10.0 was recently released, following the release last month
of RHEL (Red Hat Enterprise Linux) 10.0, and because we anticipate making
a v3.7.0 release of the Git LFS client in the near future, we added a
Dockerfile we can use to build packages for the RHEL/Rocky Linux 10
platform in commit
git-lfs/build-dockers@cc16329409175a2120e7306499b30304689b07f8 of
PR git-lfs/build-dockers#71.

We can now also add entries for this platform to the list of our
supported Linux distribution versions in the DistroMap Ruby class
source file, which is utilized by several scripts run by our
GitHub Actions release workflow jobs when building and naming the
Linux packages we publish on Packagecloud.