Makefile: correctly escape system linker flags
In commit e828a9510d054534eab37302774ee7572b44fdeb of PR #3444 we
updated our Makefile so it would pass any linker flags defined in the
LDFLAGS environment variable to the Go linker to be passed on to
the external system linker.

To do this we define the EXTRA_LD_FLAGS Makefile variable, if it is not
already defined, with the Go linker's -extldflags option and the value
of the LDFLAGS environment variable as the parameter to that option.
The EXTRA_LD_FLAGS variable is then included in the LD_FLAGS Makefile
variable, which is supplied as the parameter value of the Go compiler
command's -ldflags option in the definition of our BUILD Makefile
variable.

The "go build" compiler command passes the value of its -ldflags
option to the "go link" command, which in turn should pass the value
of its -extldflags option to the external linker.  At present, we
enclose the entire value of the -ldflags option in quotation marks,
so the full set of arguments from our LD_FLAGS variable is correctly
passed to the "go build" command as a single parameter, and then
expanded into a set of arguments for the "go link" command.

However, we do not enclose the value of the LDFLAGS environment
variable in quotation marks, so if it contains multiple arguments
for the external linker, only the first is treated as the parameter
of the -extldflags option and the remainder are incorrectly considered
to be options to the "go link" command.

In PR git-lfs/build-dockers#71 and in subsequent commits in this PR
we expect to introduce support for the recently-released Red Hat
Enterprise Linux (RHEL) 10 and Rocky Linux 10 platform, so that our
CI and release GitHub Actions workflows will build RPM packages
for this platform.

However, our builds of the Git LFS client will not succeed on this
platform unless we revise our handling of the LDFLAGS environment
variable first, because that environment variable is defined on
RHEL/Rocky Linux 10 systems with multiple options, some of which are
not valid options for the Go linker.  As a result, the "go link"
command will fail with an error such as "flag provided but not
defined: -Wl,--as-needed" when it tries to parse a -Wl option
intended for the external linker as if it were an argument to the
Go linker.

We therefore update our Makefile's definition of the EXTRA_LD_FLAGS
variable so that it enclose any value from the LDFLAGS environment
variable in escaped quotation marks.  By escaping these delimiters
we ensure they will be retained as literals within the single
parameter passed for the -ldflags option and not terminate that
parameter prematurely.  Instead, the value they enclose will now
be passed as single parameter for the -extldflags option.

We also take the opportunity to replace the curly braces around
the LDFLAGS variable name with parentheses, which is the newer
style of Makefile variable reference.

Note that, in practice, the value we pass to the -extldflags option
will be ignored because the Go compiler command does not need to invoke
an external linker, since our code no longer imports the "C" pseudo-
module nor makes any direct calls to C functions.  Our last remaining
C code was removed in commit c65ea86530141a881bea0b071c977947030af6a1
of PR #4387.  However, for the moment we retain support in our Makefile
for the use of the -extldflags option with any system-defined linker
arguments in case we have need to resort to the use of C code again
in the future.