rpm/SPECS/git-lfs.spec: drop RHEL 5 cgo workaround
In commit 6006acc68161dbe757d3d6f3fb63259013deb198 in PR #976 we
introduced a conditional build clause to the SPEC file use we to
build RPM Linux packages, in an attempt to avoid compiling with
cgo enabled on Red Hat Enterprise Linux (RHEL) 5 and CentOS 5
platforms.  This change was made after the use of the "C" pseudo-
module was introduced into the Git LFS client code in commit
ea16fd51cfe4f5deaf79d941b9edf6e15dc2bc50 of PR #952, as part of
our initial support for copy-on-write file creation.

In practice, though, this conditional build clause was ineffective,
according to the notes in PR #976.  Instead, one of the Dockerfiles
used to build RPM packages for the RHEL/CentOS 5 platform was updated
in commit git-lfs/build-dockers@39cec3667e6b41fbed725fca8ace85f361c2e80b
of PR git-lfs/build-dockers#1 to set the CGO_ENABLED environment
variable with a value of "0".

The conditional build clause in our SPEC file was never removed
afterwards, however.  Later, the Dockerfile in which the CGO_ENABLED
environment variable was set was removed in commit
git-lfs/build-dockers@d4c2fe6f797decb6862072f0dc266d92b300e51a,
and the primary Dockerfile used for building packages for the
RHEL/CentOS 5 platform was then dropped in commit
git-lfs/build-dockers@64a3a9fc4df3ca28920a4051435d55487832c5b7
of PR #3.  This followed the removal of the references to the
RHEL/CentOS 5 platform from the scripts in our "dockers" directory in
commit b560b853a6e557195c0c68875218f45c89dce56b of PR #1298.

As there is no value to retaining the conditional build clause
for the RHEL/CentOS 5 platform in the SPEC file for our RPM packages,
we simply remove it now.