t/t-{checkout,pull}.sh: indent pushd/popd blocks
In PR #527 we introduced the "git lfs checkout" and "git lfs pull"
commands and added some initial tests of those commands to our
test suite, which we have subsequently expanded over time.  Before
we adjust how these commands check and create files, we want to
further revise and expand their test suites to validate a broader
range of conditions.

As a first step, we reformat the initial "checkout" and "pull" tests
in the t/t-checkout.sh and t/t-pull.sh test scripts, specifically the
blocks where these tests perform checks after changing the current
working directory.

In the case of the "checkout" test in the t/t-checkout.sh test script,
we use the "pushd" and "popd" shell built-ins to change the current
working directory and then reverse that change.  However, the
intervening commands are not indented, so we do that now to bring the
formatting of this test in line with many of our other tests which
also use these shell built-ins.

In order to improve the consistency between our tests, we then update
the "pull" test in the t/t-pull.sh test script to also use the "pushd"
and "popd" shell built-ins around the "git lfs pull" commands we
want to run in a subdirectory of the test repository's working tree.

We also take the opportunity to remove several unnecessary ls(1) commands
which are left over from development of the tests.