t: add shell tests with pointer extension program
In PR #486 we introduced support for Git LFS pointer extensions, along
with some related tests in our Go test suite and the t/t-ext.sh shell
test script, which validates that pointer extensions are reported as
expected by the "git lfs ext" command.  (Note that Git LFS pointer
extension support is still technically considered experimental,
according to our documentation.)

However, we do not have any tests which exercise a complete Git LFS
pointer extension configuration including an extension program which
transforms the content of files tracked as Git LFS objects.

We therefore add a new lfstest-caseinverterextension test utility
program and make use of it in a set of shell tests of our "git lfs clean",
"git lfs smudge", and "git lfs filter-process" commands, as well as
in tests of our "git lfs checkout" and "git lfs pull" commands.  All
of these commands execute the lfstest-caseinverterextension utility
and confirm that it runs as expected.

Note that at present, when we execute a pointer extension program within
one of our "git lfs clean", "git lfs smudge", and "git lfs filter-process"
commands, our commands have typically inherited their execution
environment from Git.  When Git runs one of these commands to perform
a "clean" or "smudge" filter operation, it has already changed the current
working directory to be the root of the repository's working tree.  The
setup_git_directory_gently() function in Git is normally run shortly
after starting, and if it detects that the current working directory
is within a Git work tree, it changes the current working directory to
the root of that work tree:

  https://github.com/git/git/blob/v2.50.1/setup.c#L1758-L1760

The file paths Git then passes to our filter commands are always relative
to the root of the repository.  Git passes these file paths in place of
the "%f" command-line specifier from the filter configuration options, and
for long-running filter processes such as our "git lfs filter-process"
command, Git passes an equivalent file path as the value of the "pathname"
keys it sends to the filter process.  The gitattributes(5) manual page
notes that files may not actually exist at these file paths, or may have
different contents than the ones Git pipes to the filter process, and so
filter programs should not attempt to access files at these paths:

  https://github.com/git/git/blob/v2.50.1/Documentation/gitattributes.adoc?plain=1#L503-L507

Likewise, when the Git LFS client invokes a Git LFS pointer extension
program, it passes a file path in place of any "%f" command-line
specifier in the extension configuration options, while piping the
actual file contents to the program on the standard input file
descriptor.  If the client itself has been run by Git as a filter program,
then the file path will be relative to the root of the repository, and
the current working directory will be the root of the repository's
working tree.

When the Git LFS client is not run by Git as a filter program but executed
directly via the "git lfs checkout" or "git lfs pull" commands, however,
we do not change the current working directory before invoking pointer
extension programs, and pass file paths that are relative to the current
working directory rather than the repository root.  One exception is
when the --to option of the "git lfs checkout" command is specified,
in which case we pass the file path argument of that option to any
pointer extension programs instead of a path to the pointer file whose
contents they are processing.

Like Git filter programs, Git LFS pointer extension program should not
expect to access an actual file at the paths passed in place of the "%f"
command-line specifiers.  At present, though, we do not make this explicit
in our documentation.

In our new tests of the pointer extension support of our "git lfs checkout"
and "git lfs pull" commands, we specifically check that the file paths
logged by the lfstest-caseinverterextension test utility are relative
to the current working directory, except when the --to option of the
"git lfs checkout" command is specified, in which case that option's
argument is passed to the test utility.  We include comments noting that
both of these behaviours differ from that seen when an extension program
is run by a Git LFS filter command executed by Git.

In subsequent commits we expect to adjust how our "git lfs checkout" and
"git lfs pull" commands handle file paths, including by changing the
current working directory to the root of the current working tree before
writing files into that work tree.  As a consequence, these commands will
pass only paths relative to the root of the repository to the
SmudgeToFile() method of the GitFilter structure in our "lfs" package.

As such, we will also necessarily change the file paths passed to Git LFS
pointer extensions so they are also always relative to the root of the
repository, and when we invoke pointer extension programs, the current
working directory will already be set to the root of the current work
tree.

At the same time we will then revise our "checkout: pointer extension",
"checkout: pointer extension with conflict", and "pull: pointer extension"
tests so they verify that the file paths received by the
lfstest-caseinverterextension test utility are relative to the root of
the repository, and will update the utility itself so it checks that
its current working directory is the root of a Git work tree.

Co-authored-by: Lars Schneider <larsxschneider@github.com>