commands,t: log rooted paths on checkout and pull
Our "git lfs checkout" and "git lfs pull" commands call the
DecodePointerFromFile() function from our "lfs" package to check whether
a file is present at a given path in the repository's current working
tree, and if so, whether that file contains a valid Git LFS pointer.

The DecodePointerFromFile() function is called by the Run() method of
the singleCheckout structure in our "commands" package when these
"git lfs checkout" and "git lfs pull" commands are executed.  If the
DecodePointerFromFile() function returns an error, the Run() method
behaves differently depending on the type of error.

If the error is not one for which the Run() method performs any special
action, the method logs the error with a simple "Checkout error" message,
along with a relative path to the file from the current working directory.

In subsequent commits we expect to adjust how our "git lfs checkout" and
"git lfs pull" commands handle file paths, including by changing the
current working directory to the root of the current working tree before
writing files into that work tree.  As a consequence, these commands will
pass only paths relative to the root of the repository to the
SmudgeToFile() method of the GitFilter structure in our "lfs" package.
We will also enhance these commands to test for symbolic links in the
path between the root of the work tree and a given file, and will
report these links using a new log message format.

Before we make these changes, we first revise the format of the log
messages and file paths that are output when the DecodePointerFromFile()
function returns an error which the Run() method does not handle with
any special action.  Specifically, we now report a file path which is
relative to the root of the repository and not relative to the current
working directory.

On Unix systems, these log messages are the ones reported when an
ancestor component of a file's path is found to be a file and not a
directory, and the Stat() method of the "os" package in the Go
standard library returns an ENOTDIR error number.

As noted above, in subsequent commits we will enhance the "git lfs
checkout" and "git lfs pull" commands to detect similar conditions
involving symbolic links.  We would like the log messages and file
paths output by the commands under all these types of conditions to be
as consistent as possible, which is why we first update our existing
log messages to the format we expect to use in the future.