check for file/symlink conflicts on checkout/pull
Our "git lfs checkout" and "git lfs pull" commands, at present,
follow any extant symbolic links when they populate the current working
tree with files containing the content of Git LFS objects, even if
the symbolic links point to locations outside of the working tree.
This vulnerability has been assigned the identifier CVE-2025-26625.

To partially address this vulnerability, we adjust the
DecodePointerFromBlob() function in our "lfs" package to use the Lstat()
function from the "os" package in the Go standard library instead of
the Stat() function.  This ensures that the DecodePointerFromBlob()
function checks whether an irregular file or other directory entry
already exists at the location where the "git lfs checkout" and "git lfs
pull" commands intend to create or update a file.

We then update a number of the tests that we added to the t/t-checkout.sh
and t/t-pull.sh test scripts in previous commits, and now also add
another pair of new tests to those scripts.

First, we revise the "checkout: skip directory file conflicts",
"pull: skip directory file conflicts", "checkout: skip directory symlink
conflicts", and "pull: skip directory symlink conflicts" tests so that
when they run on Unix systems, they now expect the name of the lstat(2)
system call to appear in the log messages output by the "git lfs checkout"
and "git lfs pull" commands.  Previously, these tests expected the name
of the stat(2) system call to appear in the commands' log messages.

Next, we expand and revise the "checkout: skip file symlink conflicts"
and "pull: skip file symlink conflicts" tests so they confirm that the
respective commands try to avoid writing through symbolic links which
exist in the working tree at the locations where the commands intend to
create or update files, regardless of the nature of the links' targets.
In their initial form, these tests could only check the case where the
targets of the symbolic links were directories, but now they can also
check the commands' behaviour both when the links' targets do not exist
and when the targets are files which contain Git LFS pointers identical
to those of the corresponding paths in the Git repository.  Previously,
in such cases the commands would create or update files at the locations
of the targets of the symbolic links.

We then add two new tests, named "checkout: skip case-based symlink
conflicts" and "pull: skip case-based symlink conflicts", which confirm
that the respective commands do not write through symbolic links which
exist in the working tree at the locations where the commands intend to
create or update files, after those links are created by Git due to
filename conflicts on case-insensitive filesystems.  Like the other
tests with symbolic links, we only run these new tests on Windows if
the current system supports the creation of true symbolic links.

In both our new and revised tests we run the "git lfs checkout" and
"git lfs pull" commands at several directory levels in the working tree,
in order to exercise the ability for these commands to be run in any
subdirectory, a behaviour we have supported since PR #2641.  We also
confirm that the commands do not add the paths of the symbolic links to
the Git index as they previously did because the commands assumed they
had updated regular files at those locations.

Note that while our new check in the DecodePointerFromFile() function
avoids cases where a symbolic link already exists the working tree
before we try to create or update a file at the same location, this
check does not entirely prevent TOCTOU (time-of-check/time-of-use)
races where a symbolic link might be created immediately after we check
for its existence and before we attempt to create or open a file.

In a subsequent commit we will address these concerns, at least in
part, by changing the SmudgeToFile() method of the GitFilter structure
in our "lfs" package to remove any existing file or link and always
create a new file with the O_EXCL flag.  This should help ensure we only
ever create a new file and never write through a symlink that was added
immediately after the DecodePointerFromBlob() function ran.

Finally, note that other than the "git lfs checkout" and "git lfs pull"
commands, the only other caller of the DecodePointerFromBlob() function
is the "git lfs merge-driver" command, which is guaranteed by the context
in which it runs to always open regular, temporary files created by Git.
For this reason, we do not need to expand the test suite for the
"git lfs merge-driver" command to check how it handles pre-existing
symbolic links.