t: test some symlink conflicts on checkout/pull
Our "git lfs checkout" and "git lfs pull" commands, at present,
follow any extant symbolic links when they populate the current working
tree with files containing the content of Git LFS objects, even if
the symbolic links point to locations outside of the working tree.
This vulnerability has been assigned the identifier CVE-2025-26625.

Although the "git lfs checkout" and "git lfs pull" commands currently
follow symbolic links, under certain circumstances the commands may still
report errors and avoid writing to files through the links, depending
on the nature of the links' targets.

In subsequent commits we will adjust how these commands check and create
files in order to address the vulnerability described above.  Before we
make these changes, though, we first add two pairs of new tests to
validate the limited set of conditions under which the commands already
avoid writing through symbolic links.  As we later adjust the behaviour
of the "git lfs checkout" and "git lfs pull" commands we will update and
expand our new tests in tandem to help confirm that our revisions to
the commands are effective.

The first pair of new tests we add are the "checkout: skip directory
symlink conflicts" and "pull: skip directory symlink conflicts" tests,
for the "git lfs checkout" and "git lfs pull" commands, respectively.
In their initial form, these tests check that the commands report errors
and avoid writing through symbolic links which appear in place of
directories in the paths between the root of the working tree and the
locations where the commands intend to create Git LFS files, but only
when the targets of the links either do not exist or are not directories.

The second pair of new tests we add are the "checkout: skip file symlink
conflicts" and "pull: skip file symlink conflicts" tests.  In their
initial form, these tests check that the "git lfs checkout" and "git lfs
pull" commands report errors and avoid writing through symbolic links
which exist at the locations where the commands intend to create files,
but only when the targets of the links exist and are directories rather
than regular files.

All of our new tests use the assert_clean_index() test helper function
we introduced in a previous commit to confirm that in the specific
circumstances described above, the "git lfs checkout" and "git lfs pull"
commands do not update the Git index entries for the Git LFS files
whose paths now contain symbolic links in the working tree.

Note that we only run our new tests on Windows if the current system
supports the creation of true symbolic links, which we determine using
a new skip_if_symlinks_unsupported() test helper function that simply
terminates the calling tests with a zero (i.e., successful) exit code
unless the has_native_symlinks() test helper function returns a zero
exit code.  The has_native_symlinks() function, which we added in a prior
commit, checks the results of a query made with the Windows "fsutil
reparsepoint" command after the appropriate MSYS2 and Cygwin environment
variables are set to try to enable true symbolic link support in Windows.